KAT-164 allows writing to file through the api
