+++ src/main/java/health/medunited/architecture/jaxrs/ConnectorApplication.java
@@ -6,13 +6,7 @@ import java.util.Set;
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
 
-import health.medunited.architecture.jaxrs.resource.Card;
-import health.medunited.architecture.jaxrs.resource.Certificate;
-import health.medunited.architecture.jaxrs.resource.ConnectorManagement;
-import health.medunited.architecture.jaxrs.resource.ConnectorSds;
-import health.medunited.architecture.jaxrs.resource.Event;
-import health.medunited.architecture.jaxrs.resource.Metrics;
-import health.medunited.architecture.jaxrs.resource.Scheduler;
+import health.medunited.architecture.jaxrs.resource.*;
 
 @ApplicationPath("/connector")
 public class ConnectorApplication extends Application {
@@ -26,6 +20,7 @@ public class ConnectorApplication extends Application {
 		set.add(Card.class);
 		set.add(ConnectorSds.class);
 		set.add(ConnectorManagement.class);
+		set.add(Monitoring.class);
 		return set;
 	}
 }
